requirements.txt
flask
datetime
bcrypt
mysql.connector
sqlalchemy
mysqlclient
mysql-connector-python
pymongo
bson
python-dotenv